# docs/swagger.yaml
basePath: /v1
definitions:
  httpx.DataJsonResponse:
    properties:
      data: {}
    type: object
  httpx.ErrorJsonResponse:
    properties:
      error:
        type: string
    type: object
  httpx.MessageJsonResponse:
    properties:
      message:
        type: string
    type: object
...
  schemas.AhpRecommendation:
    properties:
      consistency_ratio:
        type: number
      result:
        $ref: '#/definitions/schemas.RecommendationResult'
    type: object
...
  schemas.DeleteRecommendationRequest:
    properties:
      student_id:
        type: string
    type: object
...
  schemas.Recommendation:
    properties:
      ahp:
        $ref: '#/definitions/schemas.AhpRecommendation'
      topsis:
        $ref: '#/definitions/schemas.TopsisRecommendation'
    type: object
  schemas.RecommendationResult:
    properties:
      alternative:
        type: string
      description:
        type: string
      id:
        type: integer
      score:
        type: number
    type: object
...
  schemas.StudentRecommendation:
    properties:
      ahp_results:
        items:
          $ref: '#/definitions/schemas.RecommendationResult'
        type: array
      consistency_ratio:
        type: number
      fullname:
        type: string
      id:
        type: integer
      nisn:
        type: string
      topsis_results:
        items:
          $ref: '#/definitions/schemas.RecommendationResult'
        type: array
    type: object
...
  schemas.TopsisRecommendation:
    properties:
      id:
        type: integer
      result:
        $ref: '#/definitions/schemas.RecommendationResult'
    type: object
...
info:
  contact: {}
  description: This is a Double Track REST API
  title: Double Track API
  version: "1.0"
paths:
...
  /recommendations:
    delete:
      consumes:
      - application/json
      description: Delete recommendation for a student
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Delete student recommendation request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/schemas.DeleteRecommendationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpx.MessageJsonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Delete recommendation for a student
      tags:
      - Recommendations
      - Admin
    get:
      consumes:
      - application/json
      description: Get all students recommendations
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schemas.StudentRecommendation'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get all students recommendations
      tags:
      - Recommendations
      - Admin
  /recommendations/student:
    get:
      consumes:
      - application/json
      description: Get recommendations for a student
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.Recommendation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get recommendations for a student
      tags:
      - Recommendations
      - Student
  /recommendations/student/{studentId}:
    get:
      consumes:
      - application/json
      description: Get student recommendations details
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id student
        in: path
        name: studentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httpx.DataJsonResponse'
            - properties:
                data:
                  $ref: '#/definitions/schemas.Recommendation'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpx.ErrorJsonResponse'
      summary: Get student recommendations details
      tags:
      - Recommendations
      - Admin
...
swagger: "2.0"